+++ ci/cpu/cuspatial/build_cuspatial.sh
@@ -1,0 +1,5 @@
+set -e
+
+echo "Building cuspatial"
+conda build conda/recipes/cuspatial --python=$PYTHON
+