+++ ci/cpu/libcuspatial/build_libcuspatial.sh
@@ -1,0 +1,5 @@
+set -e
+
+echo "Building libcuspatial"
+conda build conda/recipes/libcuspatial
+