+++ cpp/tests/spatial/haversine_toy.cu
@@ -16,11 +16,11 @@
 
 #include <gtest/gtest.h>
 #include <cuspatial/haversine.hpp> 
-#include <utilities/error_utils.hpp>
+#include <utilities/legacy/error_utils.hpp>
 
-#include <tests/utilities/cudf_test_utils.cuh>
-#include <tests/utilities/column_wrapper.cuh>
-#include <tests/utilities/cudf_test_fixtures.h>
+#include <tests/utilities/legacy/cudf_test_utils.cuh>
+#include <tests/utilities/legacy/column_wrapper.cuh>
+#include <tests/utilities/legacy/cudf_test_fixtures.h>
 
 struct HaversineToy : public GdfTest 
 {